Store transitions instead of timespans
This means that this crate is basically a bunch of precomputed transition data, rather than the data needed to compute those transitions at runtime. Not only is this faster, it has a higher chance of actually being correct! (Some of the timestamps in the previous version failed to account for DST correction)