Improve lookup ergonomics
Because this crate depends on datetime explicitly, we can just use the ‘sys_timezone’ function it provides without fear that it’ll do something wrong. This means we can have the more ergonomic ‘TimeZone::get()’ and ‘TimeZone::system()’ constructors when the data crate’s trait is imported.

Also, changed the example to reflect this. It lost the ability to tell when a timezone was *found* but was *incorrect*, which will have to be added back in later.

Also amend the lookup function’s signature to work around a lifetime elision issue (zoneinfo-parse@4eac23f53c2c93c0af02cee488bb136c8cc22238).